pyproject.toml
[project]
name = "smsbot-matrix"
version = "0.1.0"
description = ""
authors = [
    {name = "meska",email = "meskatech@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx (>=0.28.1,<0.29.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
